+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     * Fix: Remove debug method from, create variations from all attributes.
     * Fix: Hide out of stock products option.
     * Fix: Adding related products to cart.
+    * Fix: Fatal error related to wrong namespace use.
 * 2.1 - 2017.04.26:
     * New: Admin notices.
     * New: My Downloads panel in my account.