+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Fix: Hide out of stock products option.
     * Fix: Adding related products to cart.
     * Fix: Fatal error related to wrong namespace use.
+    * Fix: Allow to query product tags and categories at the same time.
 * 2.1 - 2017.04.26:
     * New: Admin notices.
     * New: My Downloads panel in my account.