+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Fix: Allow to query product tags and categories at the same time.
     * Fix: Adding custom settings tab by using Integration.
     * Fix: Monthly report on dashboard should not predict future.
+    * Fix: Dynamically remove shipping from cart if shipping is not required.
 * 2.1 - 2017.04.26:
     * New: Admin notices.
     * New: My Downloads panel in my account.