+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Fix: Adding custom settings tab by using Integration.
     * Fix: Monthly report on dashboard should not predict future.
     * Fix: Dynamically remove shipping from cart if shipping is not required.
+    * Fix: Coupon usage limit.
 * 2.1 - 2017.04.26:
     * New: Admin notices.
     * New: My Downloads panel in my account.