+++ CHANGELOG.md
@@ -10,6 +10,8 @@
     * Fix: Monthly report on dashboard should not predict future.
     * Fix: Dynamically remove shipping from cart if shipping is not required.
     * Fix: Coupon usage limit.
+    * Fix: Missing tax classes options in variation form.
+    * Fix: Improved prompt box for variation bulk actions.
 * 2.1 - 2017.04.26:
     * New: Admin notices.
     * New: My Downloads panel in my account.