+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * Fix: Coupon usage limit.
     * Fix: Missing tax classes options in variation form.
     * Fix: Improved prompt box for variation bulk actions.
+    * Fix: Properly migrate product attachments.
 * 2.1 - 2017.04.26:
     * New: Admin notices.
     * New: My Downloads panel in my account.